Self Correction Multi Agent System/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "self-correcting-sql-agent"
version = "0.1.0"
description = "A self-correcting multi-agent SQL system built with LangGraph"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "typing-extensions>=4.5",
    "langchain-core>=0.2",
    "langchain-groq>=0.1",
    "langchain-openai>=0.1",
    "langgraph>=0.2",
    "groq>=0.4",
    "sqlalchemy>=2.0",
    "pymysql>=1.1",
    "PyYAML>=6.0",
    "streamlit>=1.30",
]

[project.optional-dependencies]
mysql = ["pymysql>=1.1"]
postgres = ["psycopg2-binary>=2.9"]
mssql = ["pyodbc>=5.0"]
oracle = ["cx_Oracle>=8.3"]
all-db = [
    "pymysql>=1.1",
    "psycopg2-binary>=2.9",
    "pyodbc>=5.0",
    "cx_Oracle>=8.3",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "ipykernel>=6.0",
]

[project.scripts]
sql-agent = "src.cli.interactive:interactive_loop"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
